requirements.txt
fastapi
uvicorn[standard]
websockets
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart